Move extension methods in to a separate file
Extension methods are syntactic sugar to the main API 'master' methods. Moved
externally to separate out their usage. Will maintain backward compatibility.
