+++ MimeDetective/MimeDetective.csproj
@@ -41,6 +41,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="BaseExtensions.cs" />
     <Compile Include="Extensions\ArchiveExtensions.cs" />
     <Compile Include="Extensions\DocumentExtensions.cs" />
     <Compile Include="Extensions\GraphicsExtensions.cs" />
